moved interface scanner to its own file
